fix: set default protocol to empty array
This prevents an issue on firefox and edge where the browsers will
send the value 'undefined' for the header Sec-WebSocket-Protocol. This
is different to the behavior of WebKit browsers which omit the header
if the value is undefined. Using just WebSocket(url) can also fix this
problem. This undefined header can then cause further issues if the
server sends back a response with an empty Sec-WebSocket-Protocol
header, which will cause the browser to reject the WebSocket
connection as closed (ReadyState 3).