Add exports to package.json to satisfy nodejs --experimental-module users
Fixes #206
This ensures that consumers using --experimental-modules with node 12+
will import the esm version of this package.

This package's export must a package.json with "type": "json" (or module
files must have .mjs extension)...
This package's export must also use exports field in the package.json
because node doesn't support the "module" field as an alternative to
"main".
	see https://nodejs.org/api/packages.html#conditional-exports